// src/features/EditableProfileCard/model/consts/consts.ts
// 'enum' должны быть в отдельных файлах для корректной работы 'isolatedModules'

export enum ValidateProfileError {
  INCORRECT_AGE = 'INCORRECT_AGE',
  INCORRECT_AVATAR = 'INCORRECT_AVATAR',
  INCORRECT_CITY = 'INCORRECT_CITY',
  INCORRECT_COUNTRY = 'INCORRECT_COUNTRY',
  INCORRECT_CURRENCY = 'INCORRECT_CURRENCY',
  INCORRECT_USER_DATA = 'INCORRECT_USER_DATA',
  INCORRECT_USERNAME = 'INCORRECT_USERNAME',
  NO_DATA = 'NO_DATA',
  SERVER_ERROR = 'SERVER_ERROR',
}